// src/app/models/service-fbb.ts
export class ServiceFBB {
    id: string;	
	wrId: string;	
	impactFlag: string;	
	accessNode: string;	
	pon: string;	
	splitter: string;	
	ponNo: string;	
	splitterFullname: string;	
	totalCount: string;	
	chkBfCount: string;	
	chkBfBy: string;	
	chkBfDate: string;	
	chkAtCount: string;	
	chkAtBy: string;	
	chkAtDate: string;
	onlineCount: string;
	checkFlag: string;
	selected: boolean;
	sendSmsDate: string;
  }
  